+++ resources/shaders/common.glsl
@@ -13,6 +13,7 @@ uniform float uCurvature;
 uniform float uFov;
 uniform int uFlat;
 uniform int uIsXRay;
+uniform int uIsThumbnail;
 
 uniform int uBevelEnabled;
 uniform sampler2D uBevelNormalMap;
@@ -77,6 +78,9 @@ vec3 computeCurvature(const in vec3 vertex, const in vec3 normal, const in vec3 
 }
 
 vec4 encodeFragColor(const in vec3 frag, const in float alpha) {
+    if (uIsThumbnail == 1) {
+        return vec4(linearTosRGB(frag), 1.0);
+    }
     if (uIsXRay == 1) {
         return vec4(frag * alpha, alpha);
     }